requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.10.0
